back/requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15,<3.17
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
requests>=2.32,<3.0
